+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as ChildDocumentLinks } from './ChildDocumentLinks/ChildDocumentLinks.js';
 export { default as Editor } from './Editor/Editor.js';
-export { default as Navbar } from './Navbar/Navbar.js';
+export { default as Breadcrumb } from './Breadcrumb/Breadcrumb.js';
 export { default as Sidebar } from './Sidebar/Sidebar.js';
 export { default as StyleMenu } from './StyleMenu/StyleMenu.js';
