+++ src/pages/DocumentPage.html
@@ -1,6 +1,6 @@
 <aside class="sidebar"></aside>
 <main class="main">
-  <nav class="main__navbar"></nav>
+  <nav class="main__breadcrumb"></nav>
   <article class="main__editor">
     <div class="main__title-editor" data-role="title" type="text" contenteditable="true" placeholder="제목 없음"></div>
     <div class="main__content-editor" data-role="content" contenteditable="true" placeholder="내용을 입력하세요"></div>
