+++ src/pages/constructors/breadcrumb.js
@@ -1,0 +1,9 @@
+import { Breadcrumb } from '../../components/index.js';
+
+export default function createBreadcrumb() {
+  const { $target } = this;
+
+  return new Breadcrumb({
+    $target: $target.querySelector('.main__breadcrumb'),
+  });
+}
