+++ src/pages/constructors/index.js
@@ -1,5 +1,5 @@
 export { default as createChildDocumentLinks } from './childDocumentLinks.js';
 export { default as createEditor } from './editor.js';
-export { default as createNavbar } from './navbar.js';
+export { default as createBreadcrumb } from './breadcrumb.js';
 export { default as createSidebar } from './sidebar.js';
 export { default as createStyleMenu } from './styleMenu.js';
