--- src/pages/constructors/navbar.js
@@ -1,9 +1,0 @@
-import { Navbar } from '../../components/index.js';
-
-export default function createNavbar() {
-  const { $target } = this;
-
-  return new Navbar({
-    $target: $target.querySelector('.main__navbar'),
-  });
-}
